Implement a streaming data analysis
When the analysis code was first designed, it was designed around simulations up to 4M steps. Analysis involved loading the whole simulation and caching to memory. During practical application of the code, it was found that the simulations could require as many as 250M steps to reach 10ms of simulation time. Not considering multiple simulations per NPMC run, memory limits of the compute node (96-196GB) were being exceeded. Therefore, the data must be streamed from the file and analyzed on-the-fly.